Attempt to add non-root user to docker group last
This way if Docker failed to install earlier and there is no docker
group, we will still be able to log in as the non-root user. The script
will then be able to proceed and mark the server as failed successfully,
instead of getting stuck trying to SSH in as the non-root user forever.